+++ electron.vite.config.ts
@@ -3,7 +3,6 @@ import { resolve } from 'path';
 import { defineConfig, externalizeDepsPlugin } from 'electron-vite';
 import react from '@vitejs/plugin-react';
 import tsconfigPaths from 'vite-tsconfig-paths';
-import prism from 'vite-plugin-prismjs';
 
 export default defineConfig({
   main: {
@@ -32,11 +31,6 @@ export default defineConfig({
       },
     },
     plugins: [
-      prism({
-        languages: 'all',
-        theme: 'tomorrow',
-        css: true,
-      }),
       react({
         babel: {
           plugins: ['babel-plugin-react-compiler'],