+++ src/renderer/src/components/common/DiffViewer/index.tsx
@@ -1,2 +1,3 @@
-export * from './DiffViewer';
-export * from './UDiffViewer';
+export { DiffViewer } from './DiffViewer';
+export { UDiffViewer } from './UDiffViewer';
+export { CompactDiffViewer } from './CompactDiffViewer';